requirement.txt
Package                   Version
------------------------- -----------
altair                    5.5.0
async-timeout             5.0.1
attrs                     25.3.0
bidict                    0.23.1
blinker                   1.9.0
cachetools                6.1.0
certifi                   2025.8.3
charset-normalizer        3.4.2
click                     8.2.1
et_xmlfile                2.0.0
gitdb                     4.0.12
GitPython                 3.1.45
h11                       0.16.0
idna                      3.10
Jinja2                    3.1.6
jsonschema                4.25.0
jsonschema-specifications 2025.4.1
MarkupSafe                3.0.2
narwhals                  2.0.1
numpy                     2.2.6
openpyxl                  3.1.5
packaging                 25.0
pandas                    2.3.1
pillow                    11.3.0
pip                       22.0.2
plotly                    6.2.0
protobuf                  6.31.1
pyarrow                   21.0.0
pydeck                    0.9.1
python-dateutil           2.9.0.post0
python-engineio           4.12.2
python-socketio           5.13.0
pytz                      2025.2
redis                     6.3.0
referencing               0.36.2
requests                  2.32.4
rpds-py                   0.27.0
setuptools                59.6.0
simple-websocket          1.1.0
six                       1.17.0
smmap                     5.0.2
streamlit                 1.48.0
streamlit-autorefresh     1.0.1
tenacity                  9.1.2
toml                      0.10.2
tornado                   6.5.2
typing_extensions         4.14.1
tzdata                    2025.2
urllib3                   2.5.0
watchdog                  6.0.0
websocket-client          1.8.0
wsproto                   1.2.0
